#448 fix: RecommendTrip 서비스 테스트 코드를 선호 여행지 전략패턴 테스트로 전환
